+++ src/messageExtension/templates/src/app/scripts/{messageExtensionName}/{messageExtensionClassName}Config.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
-import { Provider, Flex, Header, Checkbox, Button } from "@fluentui/react";
-import TeamsBaseComponent, { ITeamsBaseComponentProps, ITeamsBaseComponentState } from "msteams-react-base-component";
+import { Provider, Flex, Header, Checkbox, Button } from "@fluentui/react-northstar";
+import TeamsBaseComponent, { ITeamsBaseComponentState } from "msteams-react-base-component";
 import * as microsoftTeams from "@microsoft/teams-js";
 
 /**
@@ -13,7 +13,7 @@ export interface I<%=messageExtensionClassName%>ConfigState extends ITeamsBaseCo
 /**
  * Properties for the <%=messageExtensionClassName%>Config React component
  */
-export interface I<%=messageExtensionClassName%>ConfigProps extends ITeamsBaseComponentProps {
+export interface I<%=messageExtensionClassName%>ConfigProps {
 
 }
 
@@ -22,7 +22,7 @@ export interface I<%=messageExtensionClassName%>ConfigProps extends ITeamsBaseCo
  */
 export class <%=messageExtensionClassName%>Config extends TeamsBaseComponent<I<%=messageExtensionClassName%>ConfigProps, I<%=messageExtensionClassName%>ConfigState> {
 
-    public componentWillMount() {
+    public acomponentWillMount() {
         this.updateTheme(this.getQueryVariable("theme"));
         this.setState({
             onOrOff: true