+++ src/tab/TabGenerator.ts
@@ -232,7 +232,7 @@ export class TabGenerator extends Generator {
             this.fs.writeJSON(manifestPath, manifest);
 
             Yotilities.addAdditionalDeps([
-                ["@fluentui/react", "~0.45.0"],
+                ["@fluentui/react-northstar", "~0.49.0"],
                 ["react", "^16.8.6"],
                 ["react-dom", "^16.8.6"]
             ], this.fs);